* page.rb: Formatting.
* page_controller.rb: Cache 'show'.  Added 'dyanmicshow' method to get
around cache when it's needed.  Created a title_param method that
takes care of joining the title 'list' if it proves necessary.  Added
antispam for proven spammers.

* admin_controller.rb: Added config and config_update to permit
configuration of the site after the fact.

* page.rhtml: Always load the javascript files.  For the 'show'
action, load the javascript updater in order to get the login div.

* config.rhtml: Site configuration screen.

* index.rhtml: Link to site config.

* show.rhtml: Remove hr.

* lgn.rhtml: Added tiny login status page.

* create.rhtml: Use page.title.

* show.rhtml: Fiddled with title.

* routes.rb: Make title snarf all final elements in order to deal with
links like /page/show/foo/bar that links to foo/bar.



git-svn-id: http://dedawiki.rubyforge.org/svn/trunk@4 f7b17897-190c-4688-811d-8cf59fb8cab0